+++ CHANGELOG.md
@@ -3,6 +3,14 @@
   - limit exports the default object of sitemap is very minimal now
   - Sitemap constructor now uses a object for its constructor
   - Sitemap no longer accepts a single string for its url
+# 3.2.2
+  - revert https everywhere added in 3.2.0. xmlns is not url.
+  - adds alias for lastmod in the form of lastmodiso
+  - fixes bug in lastmod option for buildSitemapIndex where option would be overwritten if a lastmod option was provided with a single url
+  - fixes #201, fixes #203
+# 3.2.1
+  - no really fixes ts errors for real this time
+  - fixes #193 in PR #198
 # 3.2.0
   - fixes #192, fixes #193 typescript errors
   - correct types on player:loc and restriction:relationship types